YoutubeSummarizer/requirements.txt
accelerate==0.30.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.28.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Automat==20.2.0
aws-encryption-sdk==3.2.0
aws-shell==0.2.2
awscli==1.32.93
Babel==2.14.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
blis==0.7.11
boto3==1.34.97
botocore==1.34.118
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
cloud-init==23.4.4
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
command-not-found==0.3
confection==0.1.4
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.5
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
distro==1.9.0
distro-info==1.1+ubuntu0.2
docstring_parser==0.16
docutils==0.16
easydict==1.13
ec2-hibinit-agent==1.0.0
emoji==2.12.1
en-core-web-sm @ file:///home/ubuntu/26-crawler/en_core_web_sm-3.7.1.tar.gz
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
google-api-core==2.15.0
google-api-python-client==2.112.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-search-results==2.4.2
google-serp-api==1.0.4
googleapis-common-protos==1.62.0
googlesearch-python==1.2.4
greenlet==3.0.3
h11==0.14.0
hibagent==1.0.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1
incremental==21.3.0
instructor==1.3.2
iopath==0.1.10
ipykernel==6.29.4
ipython==8.23.0
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jiter==0.4.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.2.1
langchain-anthropic==0.1.11
langchain-aws==0.1.2
langchain-community==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchainhub==0.1.15
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.1.68
language_data==1.2.0
launchpadlib==1.10.16
layoutparser==0.3.4
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.2.2
marisa-trie==1.1.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
onnx==1.16.1
onnxruntime==1.18.0
openai==1.26.0
opencv-python==4.10.0.82
openpyxl==3.1.3
ordered-set==4.1.0
orjson==3.10.2
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.8.0
pillow==10.3.0
pillow_heif==0.16.0
platformdirs==4.2.1
portalocker==2.8.2
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.17.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==24.1.0
pyparsing==3.1.1
pypdf==4.2.0
pypdfium2==4.30.0
pyrsistent==0.18.1
pyserial==3.5
pytesseract==0.3.10
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-iso639==2024.4.27
python-json-logger==2.0.7
python-magic==0.4.24
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.0.2
rapidfuzz==3.9.3
referencing==0.35.0
regex==2024.4.28
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.4.2
scikit-posthocs==0.9.0
scipy==1.13.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
serpapi==0.1.5
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.5.6
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
ssh-import-id==5.11
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.1
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.3
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
transformers==4.40.1
triton==2.3.0
Twisted==22.1.0
typer==0.9.4
types-python-dateutil==2.9.0.20240316
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
unstructured==0.14.4
unstructured-client==0.22.0
unstructured-inference==0.7.34
unstructured.pytesseract==0.3.12
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
wadllib==1.3.6
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.16.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.17.0
zope.interface==5.4.0
